+++ README.md
@@ -3,46 +3,7 @@
 
 MVP for future major upgrade;
 
-By now only Simple reader implemented;
-
-All functionality is new and not covered with test cases, 
-use on your own risk.
-
-Legacy classes are still available, yet deprecated;
-
-New example:
-
-```java
-import exceptions.CsvReadingException;
-import reader.SimpleCsvReader;
-
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.LinkedList;
-
-public class Main {
-    public static void main(String[] args) throws Exception {
-        Path simpleCsvFile = Paths.get("/home/user/file.csv");
-        try (SimpleCsvReader reader = new SimpleCsvReader(
-                new InputStreamReader(
-                        new FileInputStream(simpleCsvFile.toFile()
-                        )
-                ), ',')) {
-            while (reader.ready()) {
-                LinkedList<String> strings = reader.readRow();
-                for (String str : strings) {
-                    System.out.println(str);
-                }
-            }
-        } catch (IOException | CsvReadingException e) {
-            e.printStackTrace();
-        }
-    }
-}
-```
+Manual will be there
 
 #### Version 1.2 Legacy classes
